% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{check_col_correlation}
\alias{check_col_correlation}
\title{Check column correlations}
\usage{
check_col_correlation(dat)
}
\arguments{
\item{dat}{data.frame}
}
\description{
Check column correlations
}
\examples{
data("pringles_fat")
data_fat = pringles_fat
h_tmp = 1.3
x = data_fat[,c("c1","c2","c3","x11","x12","x21","x22")]
check_col_correlation (dat=x)
}
