% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coating-data.R
\docType{data}
\name{coating}
\alias{coating}
\title{Photoresist-coating experiment data}
\format{data.frame}
\usage{
data(coating)
}
\description{
Photoresist-coating experiment data
}
\examples{
data(coating)
\donttest{print(coating)}
}
\references{
Cornell, J.A. and Ramsey, P.J. (1998). A Generalized mixture model for
categorized-components problems with an application to a
photoresist-coating experiment. \emph{Technometrics}, 40(1), 48-61.
(\href{https://www.tandfonline.com/doi/abs/10.1080/00401706.1998.10485481}{tandfonline})
}
\keyword{datasets}
