% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{compute_aicc}
\alias{compute_aicc}
\title{compute AICc}
\usage{
compute_aicc(rss, n, p, type = "AICc")
}
\arguments{
\item{rss}{residual sum of squares}

\item{n}{number of observation}

\item{p}{number of nonzero parameters}

\item{type}{character "AICc"}
}
\description{
compute AICc
}
\examples{
compute_aicc (rss=10, n=30, p=6, type = "AICc")
}
\references{
\href{https://stats.stackexchange.com/questions/87345/calculating-aic-by-hand-in-r/}{Calculating AIC “by hand” in R in Stack Overflow}
}
