% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/design_simplex_centroid_design_3_major_component-data.R
\docType{data}
\name{design_simplex_centroid_design_3_major_component}
\alias{design_simplex_centroid_design_3_major_component}
\title{Design points for the simplex centroid design with 3 components}
\format{data.frame}
\usage{
data(design_simplex_centroid_design_3_major_component)
}
\description{
Design points for the simplex centroid design with 3 components
}
\examples{
data(design_simplex_centroid_design_3_major_component)
\donttest{print(design_simplex_centroid_design_3_major_component)}
}
\keyword{datasets}
