% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{enlist}
\alias{enlist}
\title{Create a list}
\usage{
enlist(...)
}
\arguments{
\item{...}{object to be included as elements in the list}
}
\description{
Create a list
}
\examples{
item = c(1:10)
enlist(item)
}
