% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{expand_interactions}
\alias{expand_interactions}
\title{Expand the interaction terms for each subset group, say x11, x12, or c1, c2, c3}
\usage{
expand_interactions(dat, sel_names)
}
\arguments{
\item{dat}{data frame}

\item{sel_names}{characters}
}
\description{
Expand the interaction terms for each subset group, say x11, x12, or c1, c2, c3
}
\examples{
data("pringles_fat")
data_fat = pringles_fat
h_tmp = 1.3
x = data_fat[,c("c1","c2","c3","x11","x12","x21","x22")]
expand_interactions (dat=x, sel_names=c("c1", "c2", "c3"))
}
