% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{find_condition_num}
\alias{find_condition_num}
\title{Compute the conditional number of design matrix}
\usage{
find_condition_num(x)
}
\arguments{
\item{x}{matrix to be used in svd}
}
\description{
Compute the conditional number of design matrix
}
\examples{
data("pringles_fat")
data_fat = pringles_fat
h_tmp = 1.3
x = data_fat[,c("c1","c2","c3","x11","x12","x21","x22")]
find_condition_num (x)
}
