% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pringles_candidates2search-data.R
\docType{data}
\name{pringles_candidates2search}
\alias{pringles_candidates2search}
\title{The candidate search points in the nonlinear optimization for the optimal value
in the Pringles experiment}
\format{matrix}
\usage{
data(pringles_candidates2search)
}
\description{
The candidate search points in the nonlinear optimization for the optimal value
in the Pringles experiment
}
\examples{
data(pringles_candidates2search)
\donttest{print(pringles_candidates2search)}
}
\keyword{datasets}
