% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pringles_fat-data.R
\docType{data}
\name{pringles_fat}
\alias{pringles_fat}
\title{Pringles experiment data set with the percent of Fat as the response}
\format{data.frame}
\usage{
data(pringles_fat)
}
\description{
Pringles experiment data set with the percent of Fat as the response
}
\examples{
data(pringles_fat)
\donttest{print(pringles_fat)}
}
\references{
Kang, L., Joseph, V.R. and Brenneman, W.A. (2011). Design and modeling
strategies for mixture-of-mixtures experiments. \emph{Technometrics},
53(2), 125--36.
(\href{https://www.tandfonline.com/doi/abs/10.1198/TECH.2011.08132}{tandfonline})
}
\keyword{datasets}
