% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pringles_hardness-data.R
\docType{data}
\name{pringles_hardness}
\alias{pringles_hardness}
\title{Pringles experiment data set with the Hardness as the response}
\format{data.frame}
\usage{
data(pringles_hardness)
}
\description{
Pringles experiment data set with the Hardness as the response
}
\examples{
data(pringles_hardness)
\donttest{print(pringles_hardness)}
}
\references{
Kang, L., Joseph, V.R. and Brenneman, W.A. (2011). Design and modeling
strategies for mixture-of-mixtures experiments. \emph{Technometrics},
53(2), 125--36.
(\href{https://www.tandfonline.com/doi/abs/10.1198/TECH.2011.08132}{tandfonline})
}
\keyword{datasets}
