% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy_tools.R
\name{transitive_reduction_G}
\alias{transitive_reduction_G}
\alias{minimal_G}
\title{Transitive reduction of an attribute hierarchy}
\usage{
transitive_reduction_G(G, check_acyclic = TRUE)

minimal_G(G, check_acyclic = TRUE)
}
\arguments{
\item{G}{Binary square adjacency matrix for an attribute hierarchy.}

\item{check_acyclic}{If \code{TRUE}, stop when \code{G} contains a directed
cycle. Attribute hierarchies should be DAGs.}
}
\value{
A binary square matrix with redundant transitive edges removed.
}
\description{
Returns the minimal directed acyclic graph (DAG) that preserves the same
reachability relation among attributes. In an attribute hierarchy matrix,
\code{G[a, b] = 1} means attribute \code{a} is a prerequisite of
attribute \code{b}. If an edge can be removed without changing which
attributes are reachable from which other attributes, that edge is removed.
}
\examples{
G <- matrix(0, 4, 4)
G[1, 2] <- G[2, 3] <- G[1, 3] <- G[3, 4] <- 1
transitive_reduction_G(G)
}
